DELIVERY-34-UPDATED
Переписал компонент меню. Теперь он принимает в себя массив типа:
[
	{name: 'Имя пункта меню', func: 'функция в вашем компоненте меняющая какой компонент выводится'},
	...
]
