+++ src/components/ChiefAnalytics/Dashboard.js
@@ -1,0 +1,21 @@
+import {Paper, Typography} from "@mui/material";
+import {Box} from "@mui/system";
+
+export const Dashboard = () => {
+    return (
+        <>
+            <Paper elevation={2}>
+                <Box sx={{ p: 3 }}>
+                    <Typography variant="h6">В обработке 20 заказов</Typography>
+                    <Typography variant="h6">На линии 42 курьера</Typography>
+                    <Typography variant="h6">
+                        За сегодня доставленно 54 заказа
+                    </Typography>
+                    <Typography variant="h6">
+                        Среднее время доставки 42 минуты
+                    </Typography>
+                </Box>
+            </Paper>
+        </>
+    )
+}
